Add pertag patch - Currently needs some improvement
The pertag patch *should* allow mwfact to be split on each tag
independent of each other, however using right click drag to adjust
mwfact appears to reset all tags to the default mwfact.

This is probably related to an oversight in dragmfact and will
need to be adjusted to ensure this works.