Remove faulty bstackhoriz layout for now fixes #1
I didn't use this layout much and it is far too problematic to remain
in my config.

Further research is required to determine what went wrong there.
With too many windows open (Approximately 10+) dwm would crash with
"request code=12, error code=11" and
X would return "BadAlloc )insufficient resources for operation)

I was unable to reproduce this on a stock build of dwm 6.3 (with cfacts)
inside of a VM so it's *possibly* driver related or due to me having
multiple monitors.